# openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "ExecutionTarget": {
        "description": "Execution target.",
        "enum": [
          "pool",
          "on_demand"
        ],
        "title": "ExecutionTarget",
        "type": "string"
      },
...
      "ExperimentRunResponse": {
        "description": "Experiment run response.",
        "properties": {
          "agent_version_id": {
            "description": "Id of the agent version.",
            "format": "uuid",
            "title": "Agent Version Id",
            "type": "string"
          },
          "created": {
            "description": "Creation time.",
            "format": "date-time",
            "title": "Created",
            "type": "string"
          },
          "ended_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Execution end time.",
            "title": "Ended At"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Error message.",
            "title": "Error"
          },
          "execution_target": {
            "$ref": "#/components/schemas/ExecutionTarget"
          },
          "executor_handle": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Executor handle.",
            "title": "Executor Handle"
          },
          "experiment_id": {
            "description": "Id of the experiment.",
            "format": "uuid",
            "title": "Experiment Id",
            "type": "string"
          },
          "id": {
            "description": "Experiment run id.",
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "number": {
            "description": "Per-experiment run number.",
            "title": "Number",
            "type": "integer"
          },
          "owner_id": {
            "description": "Id of the owning account.",
            "format": "uuid",
            "title": "Owner Id",
            "type": "string"
          },
          "progress": {
            "$ref": "#/components/schemas/ExperimentRunProgress",
            "description": "Job counts by status."
          },
          "score_baselines": {
            "description": "Whether the runner also scores originals missing scores.",
            "title": "Score Baselines",
            "type": "boolean"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Execution start time.",
            "title": "Started At"
          },
          "status": {
            "$ref": "#/components/schemas/ExperimentRunStatus",
            "description": "Run status."
          },
          "summary": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Aggregate diff, written at completion.",
            "title": "Summary"
          },
          "updated": {
            "description": "Last modification time.",
            "format": "date-time",
            "title": "Updated",
            "type": "string"
          }
        },
        "required": [
          "id",
          "owner_id",
          "experiment_id",
          "number",
          "status",
          "agent_version_id",
          "score_baselines",
          "execution_target",
          "executor_handle",
          "started_at",
          "ended_at",
          "summary",
          "error",
          "progress",
          "created",
          "updated"
        ],
        "title": "ExperimentRunResponse",
        "type": "object"
      },
      "ExperimentRunStatus": {
        "description": "Experiment run status.",
        "enum": [
          "pending",
          "running",
          "canceling",
          "completed",
          "failed",
          "canceled"
        ],
        "title": "ExperimentRunStatus",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JobStatus": {
        "description": "Job status.",
        "enum": [
          "pending",
          "claimed",
          "running",
          "completed",
          "failed",
          "timed_out",
          "canceled"
        ],
        "title": "JobStatus",
        "type": "string"
      },
...
      "Page_ExperimentRunResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ExperimentRunResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "Page[ExperimentRunResponse]",
        "type": "object"
      },
      "Page_JobResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/JobResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "Page[JobResponse]",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Kitaru",
    "version": "0.21.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/experiment-runs": {
      "get": {
        "description": "List experiment runs.\n\nClients observe HTTP 200 on success, 404 when no experiment has the\nfiltered experiment id, and 422 on invalid filter or pagination\nparameters.\n\nArgs:\n    service: Experiment run service.\n    actor: Caller context.\n    experiment_id: Filter on experiment id.\n    status: Filter on run status.\n    tag: Filter on attached tag name.\n    page: Page number.\n    page_size: Page size.\n\nReturns:\n    Page of experiment runs.",
        "operationId": "list_experiment_runs_v1_experiment_runs_get",
        "parameters": [
          {
            "in": "query",
            "name": "experiment_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Experiment Id"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ExperimentRunStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "tag",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tag"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 1000,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ExperimentRunResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Experiment Runs",
        "tags": [
          "experiment-runs"
        ]
      }
    },
    "/v1/experiment-runs/{run_id}": {
      "delete": {
        "description": "Delete a terminal experiment run, including its jobs.\n\nDeletes each job's config when nothing else references it.\n\nClients observe HTTP 204 on success, 404 when no experiment run has\nthis id, and 409 when the run is not terminal.\n\nArgs:\n    run_id: Id of the experiment run.\n    service: Experiment run service.\n    actor: Caller context.",
        "operationId": "delete_experiment_run_v1_experiment_runs__run_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Experiment Run",
        "tags": [
          "experiment-runs"
        ]
      },
      "get": {
        "description": "Get an experiment run by id.\n\nClients observe HTTP 200 on success and 404 when no experiment run\nhas this id.\n\nArgs:\n    run_id: Id of the experiment run.\n    service: Experiment run service.\n    actor: Caller context.\n\nReturns:\n    Stored experiment run.",
        "operationId": "get_experiment_run_v1_experiment_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExperimentRunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Experiment Run",
        "tags": [
          "experiment-runs"
        ]
      }
    },
    "/v1/experiment-runs/{run_id}/cancel": {
      "post": {
        "description": "Cancel an experiment run.\n\nPending and claimed jobs are canceled immediately, running ones\ndrain through the heartbeat path. The run lands on canceled right away\nwhen no running job remains.\n\nClients observe HTTP 200 on success, 404 when no experiment run has\nthis id, and 409 when the run is already terminal.\n\nArgs:\n    run_id: Id of the experiment run.\n    service: Experiment run service.\n    actor: Caller context.\n\nReturns:\n    Updated experiment run.",
        "operationId": "cancel_experiment_run_v1_experiment_runs__run_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExperimentRunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Experiment Run",
        "tags": [
          "experiment-runs"
        ]
      }
    },
    "/v1/experiment-runs/{run_id}/jobs": {
      "get": {
        "description": "List the jobs of an experiment run.\n\nClients observe HTTP 200 on success, 404 when no experiment run has\nthis id, and 422 on invalid filter or pagination parameters.\n\nArgs:\n    run_id: Id of the experiment run.\n    service: Experiment run service.\n    actor: Caller context.\n    status: Filter on job status.\n    page: Page number.\n    page_size: Page size.\n\nReturns:\n    Page of jobs.",
        "operationId": "list_experiment_run_jobs_v1_experiment_runs__run_id__jobs_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/JobStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 1000,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_JobResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Experiment Run Jobs",
        "tags": [
          "experiment-runs"
        ]
      }
    },
    "/v1/experiments": {
      "get": {
        "description": "List experiments.\n\nClients observe HTTP 200 on success and 422 on invalid pagination\nparameters.\n\nArgs:\n    service: Experiment service.\n    actor: Caller context.\n    name: Filter on experiment name.\n    tag: Filter on attached tag name.\n    page: Page number.\n    page_size: Page size.\n\nReturns:\n    Page of experiments.",
        "operationId": "list_experiments_v1_experiments_get",
        "parameters": [
          {
            "in": "query",
            "name": "name",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Name"
            }
          },
          {
            "in": "query",
            "name": "tag",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tag"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 1000,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ExperimentResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Experiments",
        "tags": [
          "experiments"
        ]
      },
      "post": {
        "description": "Create an experiment over a cohort with an inline replay config.\n\nThe tool policy defaults to passthrough when omitted.\n\nClients observe HTTP 201 on success, 404 when no cohort has the\nreferenced id, 409 when the name is already registered, and 422 on\ninvalid input.\n\nArgs:\n    body: Experiment create request.\n    service: Experiment service.\n    actor: Caller context.\n\nReturns:\n    Created experiment.",
        "operationId": "create_experiment_v1_experiments_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExperimentCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExperimentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Experiment",
        "tags": [
          "experiments"
        ]
      }
    },
    "/v1/experiments/{experiment_id}": {
      "delete": {
        "description": "Delete an experiment, including its tag links.\n\nClients observe HTTP 204 on success, 404 when no experiment has this\nid, and 409 while the experiment has runs.\n\nArgs:\n    experiment_id: Id of the experiment.\n    service: Experiment service.\n    actor: Caller context.",
        "operationId": "delete_experiment_v1_experiments__experiment_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "experiment_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Experiment Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Experiment",
        "tags": [
          "experiments"
        ]
      },
      "get": {
        "description": "Get an experiment by id.\n\nClients observe HTTP 200 on success and 404 when no experiment has\nthis id.\n\nArgs:\n    experiment_id: Id of the experiment.\n    service: Experiment service.\n    actor: Caller context.\n\nReturns:\n    Stored experiment.",
        "operationId": "get_experiment_v1_experiments__experiment_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "experiment_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Experiment Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExperimentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Experiment",
        "tags": [
          "experiments"
        ]
      },
      "patch": {
        "description": "Update an experiment.\n\nName and description update on any experiment. Cohort and config\nchanges are rejected once a run exists.\n\nClients observe HTTP 200 on success, 404 when no experiment or\nreferenced cohort has this id, 409 when the new name is already\nregistered or a cohort or config change hits an experiment with runs,\nand 422 on invalid input.\n\nArgs:\n    experiment_id: Id of the experiment.\n    body: Experiment update request.\n    service: Experiment service.\n    actor: Caller context.\n\nReturns:\n    Updated experiment.",
        "operationId": "update_experiment_v1_experiments__experiment_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "experiment_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Experiment Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExperimentUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExperimentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Experiment",
        "tags": [
          "experiments"
        ]
      }
    },
    "/v1/experiments/{experiment_id}/runs": {
      "get": {
        "description": "List the runs of an experiment.\n\nClients observe HTTP 200 on success, 404 when no experiment has this\nid, and 422 on invalid pagination parameters.\n\nArgs:\n    experiment_id: Id of the experiment.\n    service: Experiment run service.\n    actor: Caller context.\n    page: Page number.\n    page_size: Page size.\n\nReturns:\n    Page of experiment runs.",
        "operationId": "list_experiment_runs_v1_experiments__experiment_id__runs_get",
        "parameters": [
          {
            "in": "path",
            "name": "experiment_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Experiment Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 1000,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ExperimentRunResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Experiment Runs",
        "tags": [
          "experiments"
        ]
      },
      "post": {
        "description": "Start an experiment run.\n\nCreates the run plus one pending job per cohort session.\n\nClients observe HTTP 201 on success, 404 when no experiment or agent\nversion has the referenced id, 409 when no runnable agent version\nresolves or an on demand run resolves to a version without an image,\nand 422 when the version belongs to another agent.\n\nArgs:\n    experiment_id: Id of the experiment.\n    body: Experiment run create request.\n    service: Experiment service.\n    actor: Caller context.\n\nReturns:\n    Created experiment run.",
        "operationId": "create_experiment_run_v1_experiments__experiment_id__runs_post",
        "parameters": [
          {
            "in": "path",
            "name": "experiment_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Experiment Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExperimentRunCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExperimentRunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Experiment Run",
        "tags": [
          "experiments"
        ]
      }
    },
...
  }
}